CHanged oxygen timer to use float, due to gradual reduction necessary during boosts

+++ Pleasant Trees/Assets/Code/OxygenTimer.cs
@@ -3,7 +3,7 @@ using System.Collections;
 
 public class OxygenTimer : MonoBehaviour {
 
-    public int TimeTillAsphyx;
+    public float TimeTillAsphyx;
